refactor(event-sorting): Moved status enum to constants folder
