+++ apps/website/src/constants/event.ts
@@ -1,0 +1,5 @@
+export enum EventStatus {
+  Ongoing = "Ongoing",
+  Finished = "Finished",
+  Upcoming = "Upcoming",
+}
