+++ apps/website/src/types/event.d.ts
@@ -1,8 +1,4 @@
-export enum EventStatus {
-  Ongoing = "Ongoing",
-  Finished = "Finished",
-  Upcoming = "Upcoming",
-}
+import { EventStatus } from "@/constants/event";
 
 export interface EventTableEntry {
   id: number;
